working on improved routing

+++ map_matching/postgres_scripts/01_create_tables.sql
@@ -5,8 +5,8 @@ DROP TABLE IF EXISTS testing.valhalla_map_match_shape CASCADE;
 CREATE TABLE testing.valhalla_map_match_shape
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     distance_m float,
     geom geometry(linestring, 4326)
 );
@@ -17,8 +17,8 @@ DROP TABLE IF EXISTS testing.valhalla_map_match_shape_non_pii CASCADE;
 CREATE TABLE testing.valhalla_map_match_shape_non_pii
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     distance_m float,
     geom geometry(linestring, 4326)
 );
@@ -30,8 +30,8 @@ DROP TABLE IF EXISTS testing.valhalla_map_match_edge CASCADE;
 CREATE TABLE testing.valhalla_map_match_edge
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     edge_index integer,
     osm_id integer,
     names text[],
@@ -48,8 +48,8 @@ DROP TABLE IF EXISTS testing.valhalla_map_match_point CASCADE;
 CREATE TABLE testing.valhalla_map_match_point
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     point_index integer,
     distance_from_trace_point float,
     distance_along_edge float,
@@ -68,8 +68,8 @@ DROP TABLE IF EXISTS testing.valhalla_map_match_fail CASCADE;
 CREATE TABLE testing.valhalla_map_match_fail
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     error_code smallint,
     error text,
     http_status text,
@@ -84,8 +84,8 @@ DROP TABLE IF EXISTS testing.valhalla_route_shape CASCADE;
 CREATE TABLE testing.valhalla_route_shape
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     segment_index integer,
     distance_m float,
     point_count integer,
@@ -99,8 +99,8 @@ DROP TABLE IF EXISTS testing.valhalla_route_fail CASCADE;
 CREATE TABLE testing.valhalla_route_fail
 (
     trip_id text,
-    search_radius smallint,
-    gps_accuracy smallint,
+    search_radius double precision,
+    gps_accuracy double precision,
     segment_index integer,
     error_code smallint,
     error text,
